# public/docs/api/v1/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Web24 API",
        "description": "API for managing companies and employees",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://web24api.devel/api/v1",
            "description": "Development API Server"
        }
    ],
    "paths": {
        "/api/v1/login": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login and get authentication token",
                "operationId": "3dbb325e0dab1fa294c929506bf9c85a",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "Token-Based Authentication": []
                    }
                ]
            }
        },
        "/api/v1/companies": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get a list of companies",
                "operationId": "707a45d5190cbbe1b43f69f6e5d0131d",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "Token-Based Authentication": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Create a new company",
                "operationId": "8c4a1d582268449c242009ff20c2dfe7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCompanyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "Token-Based Authentication": []
                    }
                ]
            }
        },
        "/api/v1/companies/{id}": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get a single company by ID",
                "operationId": "f0f1a3d86918a593a0aaa37a0c9fb1c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Company ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "Token-Based Authentication": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Companies"
                ],
                "summary": "Update an existing company",
                "operationId": "e1521de8d65f0ab70f079174789b8b9c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Company ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCompanyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompanyResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "Token-Based Authentication": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Companies"
                ],
                "summary": "Delete a company",
                "operationId": "4b2b739cc0dc94bdc5feebb74e4e9914",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Company ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "Token-Based Authentication": []
                    }
                ]
            }
        },
        "/api/v1/employees": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get a list of employees",
                "operationId": "326f2e86fa40a9ba88b1cb5f2d4b62b8",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmployeeResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "Token-Based Authentication": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Create a new employee",
                "operationId": "4a10b3ada27736863f2586c56b9c19ec",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreEmployeeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmployeeResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "Token-Based Authentication": []
                    }
                ]
            }
        },
        "/api/v1/employees/{id}": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get a single employee by ID",
                "operationId": "b7cb6100cedcac60d8c23d3f9d7c3251",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Employee ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmployeeResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "Token-Based Authentication": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Employees"
                ],
                "summary": "Update an existing employee",
                "operationId": "bcdffcad24dd066f489c0e6ed83861a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Employee ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateEmployeeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmployeeResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "Token-Based Authentication": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Employees"
                ],
                "summary": "Delete an employee",
                "operationId": "43aa0966dc246d0b0373a7db65a6bf78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Employee ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "Token-Based Authentication": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "StoreCompanyRequest": {
                "title": "Store Company Request",
                "required": [
                    "name",
                    "tax_number",
                    "address"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "tax_number": {
                        "description": "1 to 13 digits",
                        "type": "string",
                        "maxLength": 13,
                        "minLength": 1
                    },
                    "address": {
                        "description": "A valid address object",
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "StoreEmployeeRequest": {
                "title": "Store Employee Request",
                "required": [
                    "name",
                    "last_name",
                    "email",
                    "address"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "last_name": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 255
                    },
                    "phone_number": {
                        "type": "string",
                        "maxLength": 20
                    },
                    "address": {
                        "description": "A valid address object",
                        "type": "object"
                    },
                    "company_ids": {
                        "description": "A valid company IDs array",
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                },
                "type": "object"
            },
            "UpdateCompanyRequest": {
                "title": "Update Company Request",
                "required": [
                    "name",
                    "tax_number",
                    "address"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "tax_number": {
                        "description": "1 to 13 digits",
                        "type": "string",
                        "maxLength": 13,
                        "minLength": 1
                    },
                    "address": {
                        "description": "A valid address object",
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UpdateEmployeeRequest": {
                "title": "Update Employee Request",
                "required": [
                    "name",
                    "last_name",
                    "email",
                    "address"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "last_name": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 255
                    },
                    "phone_number": {
                        "type": "string",
                        "maxLength": 20
                    },
                    "address": {
                        "description": "A valid address object",
                        "type": "object"
                    },
                    "company_ids": {
                        "description": "A valid company IDs array",
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                },
                "type": "object"
            },
            "AddressResource": {
                "properties": {
                    "street": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "country_code": {
                        "type": "string"
                    },
                    "postal_code": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "AuthResource": {
                "properties": {
                    "access_token": {
                        "description": "The access token used for authentication.",
                        "type": "string"
                    },
                    "token_type": {
                        "description": "The type of token, typically \"Bearer\".",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "CompanyResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "name": {
                        "type": "string"
                    },
                    "tax_number": {
                        "type": "string"
                    },
                    "address": {
                        "$ref": "#/components/schemas/AddressResource"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "employees": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EmployeeResource"
                        }
                    }
                },
                "type": "object"
            },
            "EmployeeResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "phone_number": {
                        "type": "string"
                    },
                    "address": {
                        "$ref": "#/components/schemas/AddressResource"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "companies": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CompanyResource"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "Token-Based Authentication": {
                "type": "http",
                "description": "Bearer token authentication via Sanctum",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "API Endpoints login and authentication"
        },
        {
            "name": "Companies",
            "description": "API Endpoints for Companies"
        },
        {
            "name": "Employees",
            "description": "API Endpoints for Employees"
        }
    ]
}